Creates a DataSourceMetrics component that's rendered on the datasource/[did] page.
